moving project to srx directory
